Provided the core functionality under behave/contrib/
