+++ features/environment.py
@@ -3,11 +3,8 @@
 
 from __future__ import absolute_import, print_function
 
-import copy
-import csv
-import os
-
 from behave.model import ScenarioOutline
+from behave.contrib.csv_table_from_file import load_dynamic_examples_from_csv
 
 from behave4cmd0.setup_command_shell import setup_command_shell_processors4behave
 from behave import fixture
@@ -70,20 +67,6 @@ def before_all(context):
     setup_context_with_global_params_test(context)
     setup_command_shell_processors4behave()
     print_active_tags_summary()
-
-
-def load_dynamic_examples_from_csv(example, file_path):
-    if os.path.exists(file_path):
-        orig = copy.deepcopy(example.table.rows[0])  # Make a deep copy of the original header row
-        example.table.rows = []  # Clear existing rows
-        with open(file_path, 'r') as csv_file:
-            csv_reader = csv.reader(csv_file)
-            for row in csv_reader:
-                new_row = copy.deepcopy(orig)
-                new_row.cells = [str(cell) for cell in row]
-                example.table.rows.append(new_row)
-    else:
-        raise FileNotFoundError(f"CSV file not found: {file_path}")
 
 
 def before_feature(context, feature):
